backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydub

# pyannote-audio deps (installed --no-deps to avoid torch conflicts)
# torch, torchvision, torchaudio already in NeMo container
asteroid-filterbanks
einops
huggingface-hub
lightning
pyannote-core
pyannote-database
pyannote-metrics
pyannote-pipeline
pytorch-metric-learning
rich
safetensors
soundfile
# torch-audiomentations: installed separately with --no-deps (depends on torchaudio)
# torchmetrics: installed separately with --no-deps (backtrack loop with torch pin)
